add placeholder for administrative region
extract getCountry function into shared file
