+++ public/countries.js
@@ -1,14 +1,17 @@
-export default [
-  {
+export default {
+  ca: {
     i18nString: 'ftm-canada',
-    countryCode: 'ca'
+    countryCode: 'ca',
+    administrativeRegionI18nString: 'ftm-administrative-region-province'
   },
-  {
+  fr: {
     i18nString: 'ftm-france',
-    countryCode: 'fr'
+    countryCode: 'fr',
+    administrativeRegionI18nString: 'ftm-administrative-region-department'
   },
-  {
+  us: {
     i18nString: 'ftm-united-states',
-    countryCode: 'us'
+    countryCode: 'us',
+    administrativeRegionI18nString: 'ftm-administrative-region-state'
   }
-];
+}
