+++ public/donation-form-bounce.html
@@ -2,7 +2,7 @@
 <head>
 <script>
 
-// TODO(ajwong): This is copied from location-list-map.html. Careful.
+// TODO(ajwong): This is copied from getCountry.js. Careful.
 function GetCountry() {
   const url = new URL(window.location);
   const directories = url.pathname.split("/");
